uma tela de cadastro ta aí, os links tão funcionando, nada na tela de cadastro pronto ainda
