+++ encontre_duo/src/components/cadastro.js
@@ -1,0 +1,18 @@
+import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+
+export default class Cadastro extends Component {
+    render() {
+        return (
+            <div>
+                <h1>Home!</h1>
+                <div>
+                    <Link to="/">Home</Link>
+                </div>
+                
+            </div>
+            );
+    }
+
+
+}
